feat: add flake-based nix-darwin configuration
These files live at /etc/nix-darwin/.

This configuration is the result of quite a lot of digging online:
* opinionated blog posts
* nix-darwin docs
* YT videos
and represents a first working version to achieve a configuration of my
macOS machine as similar as the current one.

Approach: most of the software was handled by Homebrew. Where available,
the equivalent nix package (https://search.nixos.org) was chosen^1.
A homebrew module is used to handle formulae, casks and MacAppStore
Apps.

[^1]: assuming that opting for the non-homebrew option would not
increase the complexity significantly.